+++ desafio2/desafio2/class_Ventas.h
@@ -26,16 +26,16 @@ public:
     std::string getTipoCombustible() const;
     std::string getMetodoPago() const;
     std::string getDocumentoCliente() const;
-    double getMonto() const;
-    double getCantidadCombustible() const;
+    int getMonto() const;
+    int getCantidadCombustible() const;
 
     //Setters
     void setFechaHora(std::string newFechaHora);
     void setTipoCombustible(std::string newTipoCombustible);
     void setMetodoPago(std::string newMetodoPago);
     void setDocumentoCliente(std::string newDocumentoCliente);
-    void setMonto(double newMonto);
-    void setCantidadCombustible(double newCantidadCombustible);
+    void setMonto(int newMonto);
+    void setCantidadCombustible(int newCantidadCombustible);
 
     //metodos
     std::string calcularFechaHora();
